pull-split-preparepy (#52)
* initial cut of clear ones

* final moving of functions

* cleaning up a bit

* tests moved

* refactor to finish copying
